% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis.R
\name{As}
\alias{As}
\alias{Ad}
\title{CO2 supply and demand function (mol / m^2 s)}
\usage{
As(C_air, C_chl, g_tc, gamma_star, R_d)

Ad(gamma_star, R_d)
}
\arguments{
\item{C_air}{Atmospheric CO2 concentration in Pa of class \code{units}}

\item{C_chl}{Chloroplastic CO2 concentration in Pa of class \code{units}}

\item{g_tc}{Total conductance to CO2 in \eqn{\mu}mol CO2) / (m\eqn{^2} s Pa) of class \code{units}}

\item{gamma_star}{Chloroplastic CO2 compensation point in Pa of class \code{units}}

\item{R_d}{Mitochondrial respiration in \eqn{\mu}mol CO2) / (m\eqn{^2} s Pa) of class \code{units}}
}
\value{
Value in mol / (m^2 s) of class \code{units}
}
\description{
CO2 supply and demand function (mol / m^2 s)

Ad
}
\details{
\bold{Supply function:}
\cr
\deqn{A = g_\mathrm{tc} (C_\mathrm{air} - C_\mathrm{chl})}{A = g_tc (C_air - C_chl)}

\bold{Demand function:}
\cr
\deqn{? = A (1 - \Gamma* / C_\mathrm{chl}) - R_\mathrm{d}}{? = A (1 - \Gamma* / C_chl) - R_d}

\tabular{lllll}{
\emph{Symbol} \tab \emph{R} \tab \emph{Description} \tab \emph{Units} \tab \emph{Default}\cr
\eqn{A} \tab \code{A} \tab photosynthetic rate \tab \eqn{\mu}mol CO2 / (m^2 s) \tab calculated \cr
\eqn{g_\mathrm{tc}}{g_tc} \tab \code{g_tc} \tab total conductance to CO2 \tab (\eqn{\mu}mol CO2) / (m\eqn{^2} s Pa) \tab \link[=.get_gtc]{calculated} \cr
\eqn{C_\text{air}}{C_air} \tab \code{C_air} \tab atmospheric CO2 concentration \tab Pa \tab 41 \cr
\eqn{C_\text{chl}}{C_chl} \tab \code{C_chl} \tab chloroplastic CO2 concentration \tab Pa \tab calculated\cr
\eqn{R_\text{d}}{R_d} \tab \code{R_d} \tab Mitochondrial respiration \tab \eqn{\mu}mol CO2 / (m\eqn{^2} s) \tab 2 \cr
\eqn{\Gamma*} \tab \code{gamma_star} \tab Chloroplastic CO2 compensation point \tab Pa \tab 3.743
}
}
