% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake-par.R
\name{bake_par}
\alias{bake_par}
\title{S3 class bake_par}
\usage{
bake_par(.x)
}
\arguments{
\item{.x}{A list to be constructed into \strong{bake_par}.}
}
\value{
Constructor function for bake_par class. This function ensures that leaf
temperature gets properly "baked" into leaf parameters.
}
\description{
S3 class bake_par
}
