% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_tdl.R
\name{cal_tdl}
\alias{cal_tdl}
\title{Calculate photosynthetic discrimination for C and O isotopes.}
\usage{
cal_tdl(dat, site_line_seq_1 = c(20, 21, 22, 23, 3, 7, 8),
  site_line_seq_2 = c(20, 21, 23, 24, 3, 7, 8), licor_ref = 7,
  licor_samp = 8, L3_tot_CO2 = 365.89, L3_d13_pdb = -37.68,
  d18O_L3_raw = -20.5, L3_d18O_SMOW = 20.12655253625,
  d13pdb_mix = -40.88, d18Ovsmow_mixing_tank = 5.223019134725,
  f_other = 0.00474, Rvpdb_C = 0.0111797, f_other_o = 0.1185,
  Rvpdb_O = 0.01185, R18Osmow = 0.0020052)
}
\arguments{
\item{dat}{clean data file for TDl (.csv).}

\item{site_line_seq_1}{site seqence used in TDL. This function expect 7 site sequnece: 4 (20,21,22,23)- caliberations, 1(3)- caliberation gas,  2 (7,8)- Licor reference and sample}

\item{site_line_seq_2}{secod site seqence (if any) used in TDL. This function expect 7 site sequnece: 4 (20,21,22,24)- caliberations, 1(3)- caliberation gas,  2 (7,8)- Licor reference and sample}

\item{licor_ref}{name licor reference line number}

\item{licor_samp}{name licor sample line number}

\item{L3_tot_CO2}{Total CO2 conecntration of caliberation gas for line 3}

\item{L3_d13_pdb}{d13C (pdb)of CO2 in caliberation gas for line 3}

\item{d18O_L3_raw}{d18O (pdb) of CO2 conecntration of caliberation gas for line 3}

\item{L3_d18O_SMOW}{d18O (vsmow) of CO2 conecntration of caliberation gas for line 3}

\item{d13pdb_mix}{d13C (pdb) of CO2 conecntration of mixing gas}

\item{d18Ovsmow_mixing_tank}{d18O (vsmow) of CO2 conecntration of mixing gas}

\item{f_other}{13C fractionation asscoiated with other isotopologus species}

\item{Rvpdb_C}{the 13C/12C ratio of the reference standard, PDB}

\item{f_other_o}{18O fractionation asscoiated with other isotopologus species}

\item{Rvpdb_O}{the 18O/16O ratio of the reference standard, PDB}

\item{R18Osmow}{the 18O/16O ratio of the reference standard, SMOW}

\item{.}{ddply function parameater}
}
\value{
Lists with raw calcuations, clean output for discrimination and caliberations used in the calculations.
}
\description{
Calculate photosynthetic discrimination for C and O isotopes using 4 cycle caliberations and one licor connection.
One can define upto two different combinations of 4 cycle caliberations.
}
