% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated-parameters.R
\encoding{UTF-8}
\name{calculated-parameters}
\alias{calculated-parameters}
\alias{get_f_parameter}
\title{Get default functions for calculated parameters in \link{photosynthesis}}
\usage{
get_f_parameter(.f_name)
}
\arguments{
\item{.f_name}{character string of function}
}
\description{
Get default functions for calculated parameters in \link{photosynthesis}
}
