% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_fluxes.R
\name{canopy_fluxes}
\alias{canopy_fluxes}
\title{Canopy fluxes}
\usage{
canopy_fluxes(meteo, leaf, LAI_layer = 20, LAI_tot = 4)
}
\arguments{
\item{meteo}{List of meteorological variables as defined in FLUXNET data with TA in °C, PA in kPa, SW in W m-2 and RH in \%}

\item{leaf}{List of leaf physiological parameters}

\item{LAI_layer}{Number of LAI layers to compute photosynthesise, default = 10

This parameter will define the extinction of light into the canopy.
If LAI_layer = 1, all leaves are considered top sun exposed leaves and will receive all the light.}

\item{LAI_tot}{Total LAI at site in m2/m2}
}
\value{

}
\description{

}
\examples{

}
