% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_to_ca.R
\name{co2_to_ca}
\alias{co2_to_ca}
\title{co2_to_ca}
\usage{
co2_to_ca(co2, patm)
}
\arguments{
\item{co2}{annual atm. CO2 in ppm}

\item{patm}{atm.pressure in Pa}
}
\value{
ca in Pa
}
\description{
Converts ambient CO2 from ppm to Pa

This function is adapted from the rpmodel package, Stocker et al. 2019
}
\examples{

}
\references{
Stocker et al., 2019 <doi:10.5194/gmd-2019-200>

<https://cran.r-project.org/web/packages/rpmodel/index.html>
}
\author{
Beni Stocker
}
