% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{S3 class constants}
\usage{
constants(.x, use_tealeaves)
}
\arguments{
\item{.x}{A list to be constructed into \strong{constants}.}

\item{use_tealeaves}{Logical. Should leaf energy balance be used to calculate leaf temperature (T_leaf)? If TRUE, \code{\link[tealeaves:tleaves]{tleaf}} calculates T_leaf. If FALSE, user-defined T_leaf is used. Additional parameters and constants are required, see \code{\link{make_parameters}}.}
}
\value{
Constructor function for constants class. This function ensures that
physical constant inputs are properly formatted.
}
\description{
S3 class constants
}
