% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductance.R
\name{.get_Dx}
\alias{.get_Dx}
\title{D_x: Calculate diffusion coefficient for a given temperature and pressure}
\usage{
.get_Dx(D_0, Temp, eT, P)
}
\arguments{
\item{D_0}{Diffusion coefficient at 273.15 K (0 degree C) and 101.3246 kPa}

\item{Temp}{Temperature in Kelvin}

\item{eT}{Exponent for temperature dependence of diffusion}

\item{P}{Atmospheric pressure in kPa}
}
\value{
Value in m\eqn{^2}/s of class \code{units}
}
\description{
D_x: Calculate diffusion coefficient for a given temperature and pressure
}
\details{
\deqn{D = D_\mathrm{0} (T / 273.15) ^ {eT} (101.3246 / P)}{D = D_0 [(T / 273.15) ^ eT] (101.3246 / P)}
\cr
According to Montieth & Unger (2013), eT is generally between 1.5 and 2. Their data in Appendix 3 indicate \eqn{eT = 1.75} is reasonble for environmental physics.
}
\references{
Monteith JL, Unsworth MH. 2013. Principles of Environmental Physics. 4th edition. Academic Press, London.
}
