% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis.R
\name{.get_gtc}
\alias{.get_gtc}
\title{Total conductance to CO2 (umol / (m^2 s Pa))}
\usage{
.get_gtc(pars)
}
\arguments{
\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\value{
Value in umol / (m^2 s Pa) of class \code{units}
}
\description{
Total conductance to CO2 (umol / (m^2 s Pa))
}
