% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis-2d.R
\name{.get_kHc}
\alias{.get_kHc}
\title{Calculate the Henryʻs Law constant for the solubility of CO2 in water as a function of temperature}
\usage{
.get_kHc(kHc25, dkHc, T_K, unitless)
}
\arguments{
\item{kHc25}{Henry's Law constant for the solubility of CO2 in water}

\item{dkHc}{Coefficient describing the temperature dependence of Henryʻs Law constant for the solubility of CO2 in water \link{K}}

\item{T_K}{Temperature \link{K}}

\item{unitless}{Logical. Should \code{units} be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.}
}
\description{
Calculate the Henryʻs Law constant for the solubility of CO2 in water as a function of temperature
}
\details{

}
\references{
Sander, R. 2015. Compilation of Henry's law constants (version 4.0) for water
as solvent. Atmos. Chem. Phys. 15: 4399–4981. https://doi.org/10.5194/acp-15-4399-2015
}
