% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductance.R
\name{.get_re}
\alias{.get_re}
\title{Re: Reynolds number}
\usage{
.get_re(pars, unitless)
}
\arguments{
\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}

\item{unitless}{Logical. Should \code{units} be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.}
}
\value{
A unitless number of class \code{units}
}
\description{
Re: Reynolds number
}
\details{
\deqn{Re = u d / D_\mathrm{m}}{Re = u d / D_m}

\tabular{lllll}{
\emph{Symbol} \tab \emph{R} \tab \emph{Description} \tab \emph{Units} \tab \emph{Default}\cr
\eqn{d} \tab \code{leafsize} \tab leaf characteristic dimension in meters \tab m \tab 0.1\cr
\eqn{D_\mathrm{m}}{D_m} \tab \code{D_m} \tab diffusion coefficient of momentum in air \tab m\eqn{^2} / s \tab \link[=.get_Dx]{calculated}\cr
\eqn{u} \tab \code{wind} \tab windspeed \tab m / s \tab 2
}
}
