% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductance.R
\name{.get_sh}
\alias{.get_sh}
\title{Sh: Sherwood number}
\usage{
.get_sh(pars, surface)
}
\arguments{
\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}

\item{surface}{Leaf surface (lower or upper)}
}
\value{
A unitless number of class \code{units}
}
\description{
Sh: Sherwood number
}
\details{
The Sherwood number depends on a combination how much free or forced convection predominates. For mixed convection: \cr 
\cr
\deqn{Sh^{3.5} = (a Re ^ b) ^ {3.5} + (c Gr ^ d) ^ {3.5})}{Sh ^ 3.5 = (a Re ^ b) ^ 3.5 + (c Gr ^ d) ^ 3.5}
\cr
\tabular{lllll}{
\emph{Symbol} \tab \emph{R} \tab \emph{Description} \tab \emph{Units} \tab \emph{Default}\cr
\eqn{a, b, c, d} \tab \code{a, b, c, d} \tab empirical coefficients \tab none \tab \link[=make_constants]{calculated}\cr
\eqn{Gr} \tab \code{Gr} \tab Grashof number \tab none \tab \link[=.get_gr]{calculated}\cr
\eqn{Re} \tab \code{Re} \tab Reynolds number \tab none \tab \link[=.get_re]{calculated}
}
}
