% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro-par.R
\name{enviro_par}
\alias{enviro_par}
\title{S3 class enviro_par}
\usage{
enviro_par(.x, use_tealeaves)
}
\arguments{
\item{.x}{A list to be constructed into \strong{enviro_par}.}

\item{use_tealeaves}{Logical. Should leaf energy balance be used to calculate leaf temperature (T_leaf)? If TRUE, \code{\link[tealeaves:tleaves]{tleaf}} calculates T_leaf. If FALSE, user-defined T_leaf is used. Additional parameters and constants are required, see \code{\link{make_parameters}}.}
}
\value{
Constructor function for enviro_par class. This function ensures that environmental parameter inputs are properly formatted.
}
\description{
S3 class enviro_par
}
