% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Licor.QC.R
\name{estimate_mode}
\alias{estimate_mode}
\title{estimate_mode}
\usage{
estimate_mode(x, adjust = 0.1)
}
\arguments{
\item{x}{numeric values}

\item{adjust}{passed to stats::density}
}
\description{
estimate_mode
}
\author{
Mike Dietze

Xiaohui Feng
}
