% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_aq_response.R
\name{fit_aq_response2}
\alias{fit_aq_response2}
\title{Fitting light responses of net CO2 assimilation}
\usage{
fit_aq_response2(
  .data,
  .vars = NULL,
  .model = "default",
  .method = "nls",
  usealpha_Q = FALSE,
  alpha_Q = 0.84
)
}
\arguments{
\item{.data}{A data frame containing CO2 assimilation light response}

\item{.vars}{A list to rename variables in .data to '.A' and '.Q' where '.A' is
net CO2 assimilation in \eqn{\mu}mol m\eqn{^{-2}} s\eqn{^{-1}} and '.Q' is
incident irradiance in \eqn{\mu}mol m\eqn{^{-2}} s\eqn{^{-1}}. '.Q' can be
corrected for light absorbance by setting \code{useapha_Q = TRUE} and setting
\code{alpha_Q = ...} based on measured or assumed leaf light absorbance.}

\item{.model}{A character string of model name to use. The default option is marshall_biscoe_1980.}

\item{.method}{A character string of the statistical method to use. Currently only "nls" (non-linear least squares) is allowed. We plan to add a Bayesion option later.}

\item{usealpha_Q}{Flag. Should light intensity be multipled by \code{alpha_Q} before fitting? Default is FALSE (i.e. assume that 'Q' is absorbed light).}

\item{alpha_Q}{Number. Absorbance of incident light. Default value is 0.84. Ignored if \code{usealpha_Q = FALSE}.}
}
\value{
\code{fit_aq_response2()} fits the light response of net CO2 assimilation and returns an \code{\link[stats]{nls}} object.
}
\description{
Fitting light responses of net CO2 assimilation
}
\note{
Rd fitted in this way is essentially the same as the Kok method, and
represents a respiration value in the light that may not be accurate.
Rd output should thus be interpreted more as a residual parameter to ensure
an accurate fit of the light response parameters. Model originally from
Marshall & Biscoe (1980).
}
\examples{
\donttest{

library(broom)
library(dplyr)
library(photosynthesis)

# Read in your data
dat = system.file("extdata", "A_Ci_Q_data_1.csv", package = "photosynthesis") |>
  read.csv() |>
  # Set grouping variable
  mutate(group = round(CO2_s, digits = 0)) |>
  # For this example, round sequentially due to CO2_s set points
  mutate(group = as.factor(round(group, digits = -1)))

# Fit one light-response curve
fit = fit_aq_response2(
  .data = filter(dat, group == 600),
  .vars = list(.A = A, .Q = Qabs)
)

# The 'fit' object inherits class 'nls' and many methods can be used

## Model summary:
summary(fit)

## Estimated parameters:
coef(fit)

## 95\% confidence intervals:
confint(fit)

## Tidy summary table using 'broom::tidy()'
tidy(fit, conf.int = TRUE, conf.level = 0.95)
}

}
\references{
Marshall B, Biscoe P. 1980. A model for C3 leaves describing the
dependence of net photosynthesis on irradiance. J Ex Bot 31:29-39
}
