% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_t_responses.R
\name{fit_t_response_arrhenius}
\alias{fit_t_response_arrhenius}
\alias{fit_t_response_heskel}
\alias{fit_t_response_kruse}
\alias{fit_t_response_medlyn}
\alias{fit_t_response_mmrt}
\alias{fit_t_response_quad}
\alias{fit_t_response_topt}
\title{Fitting temperature response functions}
\usage{
fit_t_response_arrhenius(
  data,
  varnames = list(Par = "Par", Tleaf = "Tleaf"),
  start = list(Ea = 40000, Par25 = 50),
  title = NULL
)

fit_t_response_heskel(
  data,
  varnames = list(Par = "Par", Tleaf = "Tleaf"),
  start = list(a = 1, b = 1, c = 1),
  title = NULL
)

fit_t_response_kruse(
  data,
  varnames = list(Par = "Par", Tleaf = "Tleaf"),
  start = list(dEa = 1, Ea_ref = 1, Par_ref = 1),
  title = NULL
)

fit_t_response_medlyn(
  data,
  varnames = list(Par = "Par", Tleaf = "Tleaf"),
  start = list(Ea = 40000, Hd = 2e+05, dS = 650, Par25 = mean(data$Par)),
  setvar = "none",
  Hd_set = 2e+05,
  dS_set = 650,
  title = NULL
)

fit_t_response_mmrt(
  data,
  varnames = list(Par = "Par", Tleaf = "Tleaf"),
  start = list(dCp = 1, dG = 1, dH = 1),
  title = NULL
)

fit_t_response_quad(
  data,
  varnames = list(Par = "Par", Tleaf = "Tleaf"),
  title = NULL
)

fit_t_response_topt(
  data,
  varnames = list(Par = "Par", Tleaf = "Tleaf"),
  start = list(Ea = 40000, Par25 = mean(data$Par)),
  title = NULL
)
}
\arguments{
\item{data}{Dataframe}

\item{varnames}{Variable names for the parameter to fit and leaf temperature}

\item{start}{Starting values for curve fitting (for Medlyn Arrhenius)}

\item{title}{Title for the graph}

\item{setvar}{Set a variable to constant? One of "none", "Hd" or "dS"
(for Medlyn Arrhenius)}

\item{Hd_set}{Value to set for Hd in J/mol (for Medlyn Arrhenius)}

\item{dS_set}{Value to set for dS in J/mol (for Medlyn Arrhenius)}
}
\value{
fit_t_response_arrhenius fits an Arrhenius temperature response
to data

fit_t_response_heskel fits a polynomial to log-transformed temperature
responses based on Heskel et al. 2016

fit_t_response_kruse fits the Kruse et al. 2008 model of temperature
responses

fit_t_response_medlyn fits the Medlyn et al. 2002 temperature
response model to data.

fit_t_response_mmrt fits the Macromolecular Rate Theory equation as
redefined in Liang et al. 2018, and described in Hobbs et al. 2013.

fit_t_response_quad is used to fit a 2nd order polynomial to
temperature response data

fit_t_response_topt fits the peaked Arrhenius model from
Medlyn et al. 2002 with kopt and Topt.
}
\description{
Fitting temperature response functions
}
\examples{
\donttest{
#FITTING ARRHENIUS RESPONSE CURVE
out <- fit_t_response_arrhenius(pars[pars$Plant == 1,],
varnames = list(Par = "V_cmax",
                Tleaf = "Tleaf"))
#Return summary
summary(out[[1]])

#Look at parameters
out[[2]]

#Look at graph
out[[3]]

#FITTING HESKEL RESPONSE
out <- fit_t_response_heskel(pars[pars$Plant == 1,],
varnames = list(Par = "V_cmax",
                Tleaf = "Tleaf"))
#Return summary
summary(out[[1]])

#Look at parameters
out[[2]]

#Look at graph
out[[3]]

#FITTING KRUSE RESPONSE
out <- fit_t_response_kruse(pars[pars$Plant == 1,],
varnames = list(Par = "V_cmax",
                Tleaf = "Tleaf"))
#Return summary
summary(out[[1]])

#Look at parameters
out[[2]]

#Look at graph
out[[3]]

#FITTING MEDLYN ARRHENIUS
out <- fit_t_response_medlyn(pars[pars$Plant == 1,],
varnames = list(Par = "V_cmax",
                Tleaf = "Tleaf"),
                setvar = "dS",
                dS_set = 660)
#Return summary
summary(out[[1]])

#Look at parameters
out[[2]]

#Look at graph
out[[3]]

#FITTING MMRT
out <- fit_t_response_mmrt(pars[pars$Plant == 1,],
varnames = list(Par = "V_cmax",
                Tleaf = "Tleaf"))
#Return summary
summary(out[[1]])

#Look at parameters
out[[2]]

#Look at graph
out[[3]]

#FITTING QUADRATIC RESPONSE
out <- fit_t_response_quad(pars[pars$Plant == 1,],
varnames = list(Par = "V_cmax",
                Tleaf = "Tleaf"))
#Return summary
summary(out[[1]])

#Look at parameters
out[[2]]

#Look at graph
out[[3]]

#FITTING TOPT MODEL
out <- fit_t_response_topt(pars[pars$Plant == 1,],
varnames = list(Par = "V_cmax",
                Tleaf = "Tleaf"))
#Return summary
summary(out[[1]])

#Look at parameters
out[[2]]

#Look at graph
out[[3]]
}

}
\references{
Arrhenius S. 1915. Quantitative laws in biological chemistry. Bell.

Heskel MA, O'Sullivan OS, Reich PB, Tjoelker MG, Weerasinghe LK,
Penillard A, Egerton JJG, Creek D, Bloomfield KJ, Xiang J, Sinca F,
Stangl ZR, la Torre AM, Griffin KL, Huntingford C, Hurry V, Meir P,
Turnbull MH, Atkin OK. 2016. Convergence in the temperature response
of leaf respiration across biomes and plant functional types. PNAS
113:3832-3837

Hobbs JK, Jiao W, Easter AD, Parker EJ, Schipper LA, Arcus VL.
2013. Change in heat capacity for enzyme catalysis determines
temperature dependence of enzyme catalyzed rates. ACS Chemical
Biology 8:2388-2393.

Kruse J, Adams MA. 2008. Three parameters comprehensively describe
the temperature response of respiratory oxygen reduction. Plant 
Cell Environ 31:954-967

Liang LL, Arcus VL, Heskel MA, O'Sullivan OS, Weerasinghe LK,
Creek D, Egerton JJG, Tjoelker MG, Atkin OK, Schipper LA. 2018.
Macromolecular rate theory (MMRT) provides a thermodynamics
rationale to underpin the convergent temperature response in
plant leaf respiration. Glob Chang Biol 24:1538-1547

Medlyn BE, Dreyer E, Ellsworth D, Forstreuter M, Harley PC,
Kirschbaum MUF, Le Roux X, Montpied P, Strassemeyer J, Walcroft A,
Wang K, Loutstau D. 2002. Temperature response of parameters of a
biochemically based model of photosynthesis. II. A review of
experimental data. Plant Cell Environ 25:1167-1179
}
