% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitA.R
\name{fitA}
\alias{fitA}
\title{fitA}
\usage{
fitA(flux.data, cov.data = NULL, model = NULL)
}
\arguments{
\item{flux.data}{data.frame of Licor data, concatenated by rows, and with a leading column "fname" that is used to count the number of curves and match to covariates}

\item{cov.data}{data.frame of covariate data. Column names used in formulas}

\item{model}{list including at least 6 components: the fixed effects model for alpha (a.fixed) and Vcmax (V.fixed), the random effects for these (a.random, V.random), the variable used to match the gas-exchange and covariate data (match), and the number of MCMC interations (n.iter). Additional optional arguments: TPU = TRUE turns on TPU limitation; Temp == "Bernacchi01" turns on the Bernacchi et al 2001 temperature correction. If this is turned on all parameters are estimated for 25C, otherwise no temperature correction is applied. Setting Temp = "June2004" will turn on the June et al 2004 Funct Plant Biol temperature correction to Jmax. Note: these two corrections are not mutually exclusive, you can set Temp = c("June2004","Bernacchi2001")

Right now the fixed effects are specified as a string using the standard R lm formula syntax, but without the LHS variable (e.g. "~ SLA + chl + SLA:chl"). The tilde is optional. For random effects, the two options right now are just "leaf" for leaf-level random effects and NULL. "model" has a default that sets all effects to NULL (fit one curve to all data) and n.iter=1000.}
}
\description{
fitA
}
\author{
Mike Dietze

Xiaohui Feng
}

