% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{ft}
\alias{ft}
\title{Temperature response function based on Arrhenius}
\usage{
ft(tl, ha, physcon)
}
\arguments{
\item{tl}{Leaf temperature in K}

\item{ha}{Activation energy in J mol-1}

\item{physcon}{List of physical constants}
}
\value{
This function returns the temperature correction factor f(Tl)
}
\description{
Temperature response function based on Arrhenius
}
\examples{
physcon<-load.physcon()
leaf.param<-Load.LeafPhysiologyParams()
# We compute the temperature response of Kc at tl = 25°C
ft_Kc<-ft(tl=298.15,ha=leaf.param$kcha,physcon=physcon)
}
