% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FvCB.R
\name{FvCB}
\alias{FvCB}
\alias{W_carbox}
\alias{W_regen}
\alias{W_tpu}
\title{Farquhar-von Caemmerer-Berry (FvCB) C3 photosynthesis model}
\usage{
FvCB(C_chl, pars)

W_carbox(C_chl, pars)

W_regen(C_chl, pars)

W_tpu(C_chl, pars)
}
\arguments{
\item{C_chl}{Chloroplastic CO2 concentration in Pa of class \code{units}}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\value{
A list of four values with units umol CO2 / (m^2 s) of class \code{units}:
\cr
- \code{W_carbox}: Rubisco-limited assimilation rate
- \code{W_regen}: RuBP regeneration-limited assimilation rate
- \code{W_tpu}: TPU-limited assimilation rate
- \code{A}: hyperbolic average of W_carbox, W_regen, and W_tpu
}
\description{
Farquhar-von Caemmerer-Berry (FvCB) C3 photosynthesis model

Rubisco-limited assimilation rate

RuBP regeneration-limited assimilation rate

TPU-limited assimilation rate
}
\details{
\bold{Rubisco-limited assimilation rate:} \cr
\cr
\deqn{W_\mathrm{carbox} = V_\mathrm{c,max} C_\mathrm{chl} / (C_\mathrm{chl} + K_\mathrm{m})}{W_carbox = V_cmax C_chl / (C_chl + K_m)}

where:

\deqn{K_\mathrm{m} = K_\mathrm{C} (1 + O / K_\mathrm{O})}{K_m = K_c (1 + O / K_o)}

\bold{RuBP regeneration-limited assimilation rate:} \cr
\cr
\deqn{W_\mathrm{regen} = J (C_\mathrm{chl} - \Gamma*) / (4 C_\mathrm{chl} + 8 \Gamma)}{W_regen = J C_chl / (4 C_chl + 8 \Gamma*)}

where:

\deqn{J = \theta_J PAR ^ 2 - PAR J_\mathrm{max} + PPFD \phi) + J_\mathrm{max} PPFD \phi}{J = \theta_J PAR ^ 2 - PAR J_max + PPFD \phi) + J_max PPFD \phi}

\bold{TPU-limited assimilation rate:} \cr
\cr
\deqn{W_\mathrm{tpu} = 3 V_\mathrm{tpu} C_\mathrm{chl} / (C_\mathrm{chl} - \Gamma*)}{W_tpu = 3 V_tpu C_chl / (C_chl - \Gamma*)}
}
\references{
Buckley TN and Diaz-Espejo A. 2015. Partitioning changes in photosynthetic rate into contributions from different variables. Plant, Cell & Environment 38: 1200-11.
 
Farquhar GD, Caemmerer S, Berry JA. 1980. A biochemical model of photosynthetic CO2 assimilation in leaves of C3 species. Planta 149: 78–90.
}
