% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FvCB.R
\name{J}
\alias{J}
\title{J: Rate of electron transport (mol/m^2/s)}
\usage{
J(pars)
}
\arguments{
\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\value{
Value in mol/ (m^2 s) of class \code{units}
}
\description{
J: Rate of electron transport (mol/m^2/s)
}
\details{
\eqn{J} as a function of PPFD is the solution to the quadratic expression:

\deqn{0 = \theta_J J ^ 2 - J (J_\mathrm{max} + \phi PPFD) + J_\mathrm{max} \phi PPFD}{0 = \theta_J J ^ 2 - J (J_max + \phi PPFD) + J_max \phi PPFD}
}
