% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_calculations.R
\name{calculate_jmax}
\alias{calculate_jmax}
\alias{calculate_j}
\title{Inverse non-rectangular hyperbola for J_max calculation}
\usage{
calculate_jmax(PPFD, alpha, J, theta_J)

calculate_j(PPFD, alpha, J_max, theta_J)
}
\arguments{
\item{PPFD}{light intensity in umol m-2 s-1}

\item{alpha}{initial slope of the light response}

\item{J}{electron transport rate in umol m-2 s-1}

\item{theta_J}{curvature of the light response}

\item{J_max}{maximum rate of electron transport in umol m-2 s-1}
}
\value{
calculate_jmax calculates J_max given PPFD and J.
It is necessary for the electron transport component of the
fit_aci_response function.

calculate_j provides a model of the light response of J.
It is necessary for fitting the electron transport component
of the photosynthetic CO2 response curves in fit_aci_response.
}
\description{
Inverse non-rectangular hyperbola for J_max calculation
}
