% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leakiness_high_light.R
\name{leakiness_hl}
\alias{leakiness_hl}
\title{Leakiness calculations in C4 photosynthesis under high light measurments.}
\usage{
leakiness_hl(data, varnames = c(unique_id = "unique_id", Anet = "Photo",
  Rd = "Rd", CibyCa = "Ci.Ca", gm = "gm", D13 = "D13", Ci_Pa = "Ci_Pa",
  Ca_Pa = "Ca_Pa", CL_Pa = "CL_Pa", E = "E", Cond_CO2 = "CndCO2", Tleaf =
  "Tleaf", d13_growth_air = "dgrowth_air", d13_measure_air = "d13.31",
  ...), e = 0, a = 4.4, b3 = 29, s = 1.8, ab = 2.9, am = 1.8,
  ...)
}
\arguments{
\item{data}{dataframe with variables}

\item{varnames}{List of names of variables in the dataset (see Details).}

\item{e}{CO2 fractionation during day respiration, assumed  = 0‰}

\item{a}{Weighter 12C/13 fractionation for diffusion across the boundary layer and stomata in series (4.4‰).}

\item{b3}{12C/13C fractionation during Rubisco carboxylation (29‰).}

\item{s}{12C/13C fractionation during CO2 diffusion in the boundary layer (2.9‰)}

\item{ab}{12C/13C fractionation during CO2 leackage out bundle-sheath cell assuming there is no HCO3 leackage (1.8‰).}

\item{am}{Summed 12C/13C fractionation during liquid-phase diffusion and dissolution of CO2 (1.8‰)}

\item{...}{Further arguments (ingore at the moment).}
}
\value{
A dataframe with components:
\item{t}{ternary correction factor (‰); Eqn 9}
\item{b3_bar}{(‰); Eqn 62}
\item{b4_bar}{(‰); Eqn 63}
\item{e_prime}{12C/13C fractionation during decarboxylation including the effect of
a respiratoy substrate isotipically distinct from recent photosynthate (‰); Eqn 28}
\item{leakiness}{unitless; Eqn 52 and 64}
}
\description{
Calculates leakiness under high light assuming that the,
1. bundle-sheath CO2 concentration much higher than the CO2 concentration in mesophyll cells,
2. High rate CO2 hydration i.e. Vp/Vh = 0 and Vc = A+Rd,
3. No photorespiration Vo= 0, and
4. Fractionation during day respiration = 0.

Eqn referes to Ubierna et al., 2018 (See references).
}
\details{
List of variables and their units need to suply in \code{varnames}.
\code{unique_id}{: This is a unique id for each row. Makes easy to combine output with other dataframes.}
\code{Anet}{: Net rate of CO2 assimilation; umol m^-2s^-1,}
\code{Rd}{: Rate of respiration in light; umol m^-2s^-1,}
\code{CibyC}{: Ratio of intercelluar CO2 concentration to the ambient,}
\code{gm}{: Measoplyll cell CO2 conductance' umol m^-2s^-1 Pa^-1,}
\code{D13}{: Observed 13C photsynthetic discrimination (Eqn5); ‰,}
\code{Ci_Pa}{: CO2 partial pressure inside the leaf; Pa,}
\code{Ca_Pa}{: CO2 partial pressure in the ambient air; Pa,}
\code{CL_Pa}{: CO2 partial pressure at the leaf surface; Pa,}
\code{E}{: Transpiration rate; mol m^-2s^-1,}
\code{Cond_CO2}{: Total conductance to diffusion of CO2 in air; mol m^-2 s^-1,}
\code{Tleaf}{: Leaf temperature; degree C,}
\code{d13_growth_air}{: delta13C of the CO2 at the plant growth environment,}
\code{d13_measureair}{: delta13C of the CO2 in the air used during gas-exchange measurement (Licor reference line).}
}
\references{
Ubierna N, Holloway-Phillips M-M, Farquhar GD (2018) Using Stable Carbon Isotopes to
Study C3 and C4 Photosynthesis: Models and Calculations. In S Covshoff, ed, Photosynthesis: Methods and Protocols. Springer New York, New York, NY, pp 155–196

von Caemmerer S, Ghannoum O, Pengelly JJL, Cousins AB (2014) Carbon isotope discrimination as a tool to explore C4 photosynthesis. J Exp Bot 65: 3459–3470.

Farquhar G (1983) On the nature of carbon isotope discrimination in C4 species. Functional Plant Biol 10: 205–226
}
