% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.LeafPhysiologyParams.R
\name{Load.LeafPhysiologyParams}
\alias{Load.LeafPhysiologyParams}
\title{Load leaf physiological parameters}
\usage{
Load.LeafPhysiologyParams()
}
\value{
This function return a list of leaf physiological parameters needed for the FcVB model
}
\description{
Load leaf physiological parameters
}
\examples{
leaf.param<-Load.LeafPhysiologyParams()
}
