% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{make_parameters}
\alias{make_parameters}
\alias{make_leafpar}
\alias{make_enviropar}
\alias{make_constants}
\title{Make lists of parameters for \code{photosynthesis}}
\usage{
make_leafpar(replace = NULL)

make_enviropar(replace = NULL)

make_constants(replace = NULL)
}
\arguments{
\item{replace}{A named list of parameters to replace defaults. If \code{NULL}, defaults will be used.}
}
\value{
\code{make_leafpar}: An object inheriting from class \code{\link{leaf_par}}\cr
\code{make_enviropar}: An object inheriting from class \code{\link{enviro_par}}\cr
\code{make_constants}: An object inheriting from class \code{\link{constants}}
}
\description{
Make lists of parameters for \code{photosynthesis}

make_leafpar

make_enviropar

make_constants
}
