% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_default_model}
\alias{get_default_model}
\alias{get_all_models}
\title{Get default model}
\usage{
get_default_model(method)

get_all_models(method)
}
\arguments{
\item{method}{Character string. One of "aq_response"}
}
\value{
A character string with name of model.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Get the name of the default model used for different plant ecophysiological data analysis methods implemented in \bold{photosynthesis}. Currently only used for \code{\link{fit_aq_response2}}.
}
