% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis-2d.R
\name{photo_2d_porousmedium}
\alias{photo_2d_porousmedium}
\title{Simulate C3 photosynthesis in two dimensions using a porous medium approximation}
\usage{
photo_2d_porousmedium(
  leaf_par,
  enviro_par,
  bake_par,
  constants,
  quiet = FALSE,
  assert_units = TRUE,
  check = TRUE
)
}
\description{
Simulate C3 photosynthesis in two dimensions using a porous medium approximation
}
