% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis-package.R
\docType{package}
\name{photosynthesis-package}
\alias{photosynthesis-package}
\title{\code{photosynthesis} package}
\description{
Tools for Plant Ecophysiology & Modeling
}
\details{
See the README on
\href{https://github.com/cdmuir/photosynthesis}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cdmuir/photosynthesis}
  \item Report bugs at \url{https://github.com/cdmuir/photosynthesis/issues}
}

}
\author{
\strong{Maintainer}: Chris Muir \email{cdmuir@wisc.edu} (\href{https://orcid.org/0000-0003-2555-3878}{ORCID})

Authors:
\itemize{
  \item Joseph Stinziano \email{josephstinziano@gmail.com} (\href{https://orcid.org/0000-0002-7628-4201}{ORCID})
  \item Cassaundra Roback \email{croback@unm.edu}
  \item Demi Sargent \email{d.sargent@westernsydney.edu.au}
  \item Bridget Murphy \email{bridget.murphy@mail.utoronto.ca}
  \item Patrick Hudson \email{hudson.patrick.j@gmail.com} [data contributor]
}

}
\keyword{internal}
