% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitA.R
\name{read.Licor}
\alias{read.Licor}
\title{read.Licor}
\usage{
read.Licor(filename, sep = "\\t", ...)
}
\arguments{
\item{filename}{name of the file to read}

\item{sep}{file delimiter. defaults to tab}

\item{...}{optional arguements forwarded to read.table}
}
\author{
Mike Dietze
}
