% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_piccaro.R
\name{read.picarro}
\alias{read.picarro}
\title{Reads and cleans data from PICARRO output file.}
\usage{
read.picarro(file_path)
}
\arguments{
\item{file_path}{file path for picarro .dat file.}
}
\value{
Reads and cleans data from PICARRO output file.
}
\description{
Reads and cleans data from PICARRO output file.
}
