% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tdl.R
\name{read.tdl}
\alias{read.tdl}
\title{Reads and cleans data from TDL output file (.dat).}
\usage{
read.tdl(file_path)
}
\arguments{
\item{file_path}{file path for TDl .dat file.}
}
\value{
Read, clean and compute C and O isotopic discrimination from TDL output file.
}
\description{
Read, clean and compute C and O isotopic discrimination from TDL output file.
}
