% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licor.R
\name{read_li6800}
\alias{read_li6800}
\title{Read a LI-COR file}
\usage{
read_li6800(x)
}
\arguments{
\item{x}{File name}
}
\value{
Returns a data.frame from raw LI-COR files. Current support
for LI-COR LI-6800 files only.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

We are no longer updating this function. Please use \code{\link{read_licor}} instead.
}
