% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licors.R
\name{read_li6800}
\alias{read_li6800}
\title{Reading in LiCor files}
\usage{
read_li6800(x)
}
\arguments{
\item{x}{File name}
}
\value{
Returns a dataframe from raw LiCor files. Current support
for LiCor 6800 files only. LiCor 6400 file reading will be supported
in a later version.
}
\description{
Reading in LiCor files
}
