% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{replace_defaults}
\alias{replace_defaults}
\title{Replace default parameters}
\usage{
replace_defaults(obj, replace)
}
\arguments{
\item{obj}{List of default values}

\item{replace}{List of replacement values}
}
\description{
Replace default parameters
}
