% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp-par.R
\name{temp_par}
\alias{temp_par}
\title{S3 class temp_par}
\usage{
temp_par(.x)
}
\arguments{
\item{.x}{A list to be constructed into \strong{temp_par}.}
}
\description{
Constructor function for temp_par class. This function ensures that leaf temperature parameter inputs are properly formatted.
}
