% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temper.R
\encoding{UTF-8}
\name{temper}
\alias{temper}
\alias{temp_resp1}
\alias{temp_resp2}
\title{S3 class tempered}
\usage{
temper(leaf_par, temp_par, constants)

temp_resp1(par25, E_a, R, T_leaf, T_ref)

temp_resp2(par25, D_s, E_a, E_d, R, T_leaf, T_ref)
}
\arguments{
\item{leaf_par}{A list of leaf parameters inheriting class \code{leaf_par}. This can be generated using the \code{make_leafpar} function.}

\item{temp_par}{A list of temperature response parameters inheriting class \code{temp_par}. This can be generated using the \code{make_temppar} function.}

\item{constants}{A list of physical constants inheriting class \code{constants}. This can be generated using the \code{make_constants} function.}

\item{par25}{Parameter value at 25 °C of class \code{units}.}

\item{E_a}{Empirical temperature response value in J/mol of class \code{units}.}

\item{R}{Ideal gas constant in J / (mol K) of class \code{units}. See \code{\link{make_constants}}.}

\item{T_leaf}{Leaf temperature in K of class \code{units}. Will be converted to °C.}

\item{T_ref}{Reference temperature in K of class \code{units}.}

\item{D_s}{Empirical temperature response value in J / (mol K) of class \code{units}.}

\item{E_d}{Empirical temperature response value in J/mol of class \code{units}.}
}
\description{
Constructor function for \code{tempered} class. This will also inherit class \code{\link{leaf_par}}. This function ensures that leaf parameter are "tuned" to \code{T_leaf} using temperature response functions detailed below.
}
\details{
Temperature response functions following Bernacchi et al. 2002. \cr
\cr
Temperature response function 1 (\code{temp_response1}): \cr

\deqn{par(T_\mathrm{leaf}) = par25 exp(E_\mathrm{a} / (R T_\mathrm{ref}) (T_\mathrm{leaf} - 25) / (T_\mathrm{leaf} + 273.15))}{par(T_leaf) = par25 exp(E_a / (R T_ref) (T_leaf - 25) / (T_leaf + 273.15))}

\eqn{T_\mathrm{ref}}{T_ref} is the reference temperature in K \cr
\eqn{T_\mathrm{leaf}}{T_leaf} is the leaf temperature in °C \cr
\cr
Temperature response function 2 (\code{temp_response2}) is the above equation multiplied by: \cr

\deqn{(1 + exp((D_\mathrm{s} / R - E_\mathrm{d} / (R T_\mathrm{ref})))) / (1 + exp((D_\mathrm{s} / R) - (E_\mathrm{d} / (R (T_\mathrm{leaf} + 273.15)))))}{(1 + exp((D_s / R - E_d / (R T_ref)))) / (1 + exp((D_s / R) - (E_d / (R (T_leaf + 273.15)))))}
}
\examples{
leaf_par <- make_leafpar(replace = list(T_leaf = set_units(293.15, "K")))
temp_par <- make_temppar()
constants <- make_constants()
tempered_leafpar <- temper(leaf_par, temp_par, constants)

tempered_leafpar$V_cmax25
tempered_leafpar$V_cmax

}
\references{
Bernacchi CJ, Portis AR, Nakano H, von Caemmerer S, Long SP. 2002. Temperature response of mesophyll conductance. Implications for the determination of Rubisco enzyme kinetics and for limitations to photosynthesis in vivo. Plant Physiology 130: 1992-8.
}
