% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo2_function_square.R
\name{value_square}
\alias{value_square}
\title{Squares function
This functions takes the squares of the values.}
\usage{
value_square(x)
}
\arguments{
\item{x}{numeric values to be squared}
}
\value{
square value of the input (x)
}
\description{
Squares function
This functions takes the squares of the values.
}
