% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{S3 class constants}
\usage{
constants(.x)
}
\arguments{
\item{.x}{A list to be constructed into \strong{constants}.}
}
\description{
Constructor function for constants class. This function ensures that physical constant inputs are properly formatted.
}
