% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductance.R
\name{.get_ps}
\alias{.get_ps}
\title{Saturation water vapour pressure (kPa)}
\usage{
.get_ps(Temp, P)
}
\arguments{
\item{Temp}{Temperature in Kelvin}

\item{P}{Atmospheric pressure in kPa}
}
\value{
Value in kPa of class \code{units}
}
\description{
Saturation water vapour pressure (kPa)
}
\details{
Goff-Gratch equation (see http://cires1.colorado.edu/~voemel/vp.html) \cr
\cr
This equation assumes P = 1 atm = 101.3246 kPa, otherwise boiling temperature needs to change \cr
}
\references{
\url{http://cires1.colorado.edu/~voemel/vp.html}
}
