% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro-par.R
\name{enviro_par}
\alias{enviro_par}
\title{S3 class enviro_par}
\usage{
enviro_par(.x)
}
\arguments{
\item{.x}{A list to be constructed into \strong{enviro_par}.}
}
\description{
Constructor function for enviro_par class. This function ensures that environmental parameter inputs are properly formatted.
}
