% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\encoding{UTF-8}
\name{make_parameters}
\alias{make_parameters}
\alias{make_leafpar}
\alias{make_enviropar}
\alias{make_bakepar}
\alias{make_constants}
\title{Make lists of parameters for \code{photosynthesis}}
\usage{
make_leafpar(replace = NULL, use_tealeaves)

make_enviropar(replace = NULL, use_tealeaves)

make_bakepar(replace = NULL)

make_constants(replace = NULL, use_tealeaves)
}
\arguments{
\item{replace}{A named list of parameters to replace defaults. If \code{NULL}, defaults will be used.}
}
\value{
\code{make_leafpar}: An object inheriting from class \code{\link{leaf_par}}\cr
\code{make_enviropar}: An object inheriting from class \code{\link{enviro_par}}\cr
\code{make_bakepar}: An object inheriting from class \code{\link{bake_par}}\cr
\code{make_constants}: An object inheriting from class \code{\link{constants}}
}
\description{
Make lists of parameters for \code{photosynthesis}

make_leafpar

make_enviropar

make_bakepar

make_constants
}
\examples{
leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
bake_par <- make_bakepar()
constants <- make_constants()

leaf_par <- make_leafpar(
  replace = list(
    g_sc = set_units(3, "umol/m^2/s/Pa"),
    V_cmax25 = set_units(100, "umol/m^2/s")
  )
)

}
\references{
Buckley TN and Diaz-Espejo A. 2015. Partitioning changes in photosynthetic rate into contributions from different variables. Plant, Cell & Environment 38: 1200-11.
}
