% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_names.R
\name{parameter_names}
\alias{parameter_names}
\title{Get vector of parameter names}
\usage{
parameter_names(which, use_tealeaves)
}
\arguments{
\item{which}{A character string indicating which parameter names to retreive: "leaf", "enviro", "bake", or "constants". Partial matching allowed.}

\item{use_tealeaves}{Logical. Should leaf energy balance be used to calculate leaf temperature (T_leaf)? If TRUE, \code{\link[tealeaves]{tleaf}} calculates T_leaf. If FALSE, user-defined T_leaf is used. Additional parameters and constants are required, see \code{\link{make_parameters}}.}
}
\description{
Get vector of parameter names
}
\examples{
parameter_names("leaf", use_tealeaves = FALSE)

}
