% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_names.R
\name{parameter_names}
\alias{parameter_names}
\title{Get vector of parameter names}
\usage{
parameter_names(which)
}
\arguments{
\item{which}{A character string indicating which parameter names to retreive, "leaf", "enviro", or "constants". Partial matching allowed.}
}
\description{
Get vector of parameter names
}
