% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{photo_parameters}
\alias{photo_parameters}
\title{Input parameters to simulate C3 photosynthesis using \code{\link{photosynthesis}}}
\format{
## `photo_parameters`
A data frame with `r nrow(photo_parameters)` rows and `r ncol(photo_parameters)`  columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
photo_parameters
}
\description{
A table of input parameters used in \code{\link{photosynthesis}}
}
\keyword{datasets}
