% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis-package.R
\docType{package}
\name{photosynthesis-package}
\alias{photosynthesis-package}
\title{\code{photosynthesis} package}
\description{
Tools for Plant Ecophysiology & Modeling
}
\details{
See the README on
\href{https://github.com/cdmuir/photosynthesis}{GitHub}
}
