% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis-package.R, R/photosynthesis.R
\docType{package}
\name{photosynthesis}
\alias{photosynthesis}
\alias{photosynthesis-package}
\alias{photo}
\title{\code{photosynthesis} package}
\usage{
photosynthesis(leaf_par, enviro_par, constants, progress = TRUE,
  quiet = FALSE)

photo(leaf_par, enviro_par, constants, quiet = FALSE)
}
\arguments{
\item{leaf_par}{A list of leaf parameters. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters. This can be generated using the \code{make_enviropar} function.}

\item{constants}{A list of physical constants. This can be generated using the \code{make_constants} function.}

\item{progress}{Logical. Should a progress bar be displayed?}

\item{quiet}{Logical. Should messages be displayed?}
}
\value{
\code{photosynthesis}: \cr
\cr
A data.frame (more information coming soon!)
\cr
}
\description{
Model C3 Photosynthesis
}
\details{
See the README on
\href{https://github.com/cdmuir/photosynthesis}{GitHub}
}
\examples{

leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
constants <- make_constants()
photo(leaf_par, enviro_par, constants)

}
