% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photosynthesis-package.R, R/photosynthesis.R
\docType{package}
\name{photosynthesis}
\alias{photosynthesis}
\alias{photosynthesis-package}
\alias{photo}
\title{\code{photosynthesis} package}
\usage{
photosynthesis(leaf_par, enviro_par, bake_par, constants,
  progress = TRUE, quiet = FALSE)

photo(leaf_par, enviro_par, bake_par, constants, quiet = FALSE)
}
\arguments{
\item{leaf_par}{A list of leaf parameters inheriting class \code{leaf_par}. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters inheriting class \code{enviro_par}. This can be generated using the \code{make_enviropar} function.}

\item{bake_par}{A list of temperature response parameters inheriting class \code{bake_par}. This can be generated using the \code{make_bakepar} function.}

\item{constants}{A list of physical constants inheriting class \code{constants}. This can be generated using the \code{make_constants} function.}

\item{progress}{Logical. Should a progress bar be displayed?}

\item{quiet}{Logical. Should messages be displayed?}
}
\value{
\code{photosynthesis}: \cr
\cr
A data.frame (more information coming soon!)
\cr
}
\description{
Model C3 Photosynthesis
}
\details{
See the README on
\href{https://github.com/cdmuir/photosynthesis}{GitHub}
}
\examples{
# Single parameter set with 'photo'

leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
bake_par <- make_bakepar()
constants <- make_constants()
photo(leaf_par, enviro_par, bake_par, constants)

# Multiple parameter sets with 'photosynthesis'

leaf_par <- make_leafpar(
  replace = list(
    T_leaf = set_units(c(293.14, 298.15), "K")
    )
  )
enviro_par <- make_enviropar()
bake_par <- make_bakepar()
constants <- make_constants()
photosynthesis(leaf_par, enviro_par, bake_par, constants)

}
