% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitA.R
\name{read_Licor}
\alias{read_Licor}
\title{read_Licor}
\usage{
read_Licor(filename, sep = "\\t", ...)
}
\arguments{
\item{filename}{name of the file to read}

\item{sep}{file delimiter. defaults to tab}

\item{...}{optional arguements forwarded to read.table}
}
\description{
read_Licor
}
\author{
Mike Dietze
}
