% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake.R
\name{temp_resp1}
\alias{temp_resp1}
\alias{temp_resp2}
\title{Temperature response function 1}
\usage{
temp_resp1(par25, E_a, R, T_leaf, T_ref)

temp_resp2(par25, D_s, E_a, E_d, R, T_leaf, T_ref)
}
\arguments{
\item{par25}{Parameter value at 25 °C of class \code{units}.}

\item{E_a}{Empirical temperature response value in J/mol of class \code{units}.}

\item{R}{Ideal gas constant in J / (mol K) of class \code{units}. See \code{\link{make_constants}}.}

\item{T_leaf}{Leaf temperature in K of class \code{units}. Will be converted to °C.}

\item{T_ref}{Reference temperature in K of class \code{units}.}

\item{D_s}{Empirical temperature response value in J / (mol K) of class \code{units}.}

\item{E_d}{Empirical temperature response value in J/mol of class \code{units}.}
}
\description{
Temperature response function 1

Temperature response function 2
}
