% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\usage{
test()
}
\description{
test
}
\details{
The table is:\tabular{lcr}{
   Tables \tab Are \tab Cool \cr
   col 3 is \tab \eqn{A}{A} \tab $1600 \cr
   col 2 is \tab A \tab $12 \cr
}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{# This block of code will be evaluated
tibble::tibble(
 x = c("$A$", "\\\eqn{A}{A}"),
 y = c("\\\code{A}", "25 °C")
) |>
knitr::kable()
}\if{html}{\out{</div>}}\tabular{ll}{
   x \tab y \cr
   $A$ \tab \code{A} \cr
   \eqn{A}{A} \tab 25 °C \cr
}
}
